Completada Historia de usuario 2 con un hermoso Hola Mundo
