+++ .env
@@ -1,0 +1,1 @@
+DATABASE_URL= postgres://default:jZ1GA5JOmipk@ep-tiny-sky-a4nu73yw-pooler.us-east-1.aws.neon.tech:5432/verceldb?sslmode=require
